# src/lib/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EduWeb API",
        "description": "Documentación de la API para la plataforma educativa EduWeb, diseñada para ofrecer una experiencia de aprendizaje en línea integral.",
        "version": "1.0.0"
    },
...
    "paths": {
...
        "/users/{id}": {
            "get": {
                "tags": ["Usuarios"],
                "summary": "Obtener usuario por ID",
                "description": "Recuperar un usuario específico por su ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Usuario obtenido exitosamente"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error al obtener el usuario"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": ["Usuarios"],
                "summary": "Actualizar usuario",
                "description": "Actualizar la información de un usuario existente",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "user": {
                                        "$ref": "#/components/schemas/UserInput"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Usuario actualizado exitosamente"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error al actualizar el usuario"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": ["Usuarios"],
                "summary": "Eliminar usuario",
                "description": "Eliminar un usuario del sistema",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario eliminado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Usuario eliminado exitosamente"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error al eliminar el usuario"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/users/{id}/courses/{courseId}/exams/start": {
            "get": {
                "tags": ["Cursos"],
                "summary": "Iniciar examen del curso",
                "description": "Iniciar o reanudar un examen para el curso de un usuario",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "courseId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Examen iniciado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "user_exam_id": {
                                                    "type": "string",
                                                    "format": "uuid"
                                                },
                                                "user_course_id": {
                                                    "type": "string",
                                                    "format": "uuid"
                                                },
                                                "exam_state": {
                                                    "type": "string",
                                                    "enum": ["progress", "completed"]
                                                }
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Examen iniciado exitosamente"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error al iniciar el examen"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "User": {
                "type": "object",
                "properties": {
                    "user_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "user_name": {
                        "type": "string"
                    },
                    "user_lastname": {
                        "type": "string"
                    },
                    "user_email": {
                        "type": "string",
                        "format": "email"
                    },
                    "user_phone": {
                        "type": "string"
                    },
                    "user_image": {
                        "type": "string"
                    },
                    "role": {
                        "$ref": "#/components/schemas/Role"
                    }
                }
            },
            "UserInput": {
                "type": "object",
                "properties": {
                    "user_name": {
                        "type": "string"
                    },
                    "user_lastname": {
                        "type": "string"
                    },
                    "user_email": {
                        "type": "string",
                        "format": "email"
                    },
                    "user_password": {
                        "type": "string"
                    },
                    "user_phone": {
                        "type": "string"
                    },
                    "role_id": {
                        "type": "integer"
                    }
                }
            },
            "Role": {
                "type": "object",
                "properties": {
                    "role_id": {
                        "type": "integer"
                    },
                    "role_name": {
                        "type": "string"
                    }
                }
            },
...
        }
    }
}